feat: Adding new function recarding the particle scattering for CPU and GPU
Using the `numba` package, further improvements and additions have been made to provide a speed up in the code performance. This happens on either the CPU using multithreading or on the GPU using CUDA. The code is currently split apart for the coupling matrix and scattering functionality -> merge them later on!