feat: Implement full functionality of the Optics class
The `optics` class has been implemented completely so it calculated the needed cross-sections and phase function for a particle cluster. Furthermore, the `cpu_numba.py` and `cuda_numby.py` files have been merged into a single file, so all `cpu` and `gpu` functions are in one place. This is subject to change in the future, if the files get way too bloated.