fix: mkdocs trigger change and removing mkdocs from unittest workflow
The mkdocs workflow will now only trigger if the docs folder has changed but not src or tests. The artifacts will only be downloaded, if the event trigger is a completed workflow run.
Because of the previous changes, the retention policy of the unittest workflow is changed back to 1 day, and the mkdocs step is removed from it.